refactor(codegen): extrai calls/super_calls.rs
Move 4 fns de chamadas via super (~258 LOC) para
expressions/calls/super_calls.rs:
- lower_super_call: super(args) constructor
- lower_super_prop_read / lower_super_prop_assign: super.prop
- lower_super_method_call: super.method(args)

Os 2 prop_* viraram pub(crate) para suportar re-export em
calls/mod.rs (consumido pelo lower em outros arquivos de expressions/).

calls/mod.rs: 3247 -> 2994 linhas. Suite verde.
